HOUSEKEEPING: Move feature files intro feature/ folder
