+++ src/addons/addon.cpp
@@ -25,7 +25,7 @@
 #include "conditionwatchers/addonconditionwatchertranslationthreshold.h"
 #include "conditionwatchers/addonconditionwatchertriggertimesecs.h"
 #include "env.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "glean/generated/metrics.h"
 #include "leakdetector.h"
 #include "localizer.h"
