+++ src/addons/addonapi.cpp
@@ -9,10 +9,10 @@
 
 #include "addon.h"
 #include "env.h"
+#include "feature/featuremodel.h"
 #include "frontend/navigator.h"
 #include "leakdetector.h"
 #include "logger.h"
-#include "models/featuremodel.h"
 #include "qmlengineholder.h"
 #include "settingsholder.h"
 #include "urlopener.h"
