+++ src/cmake/sources.cmake
@@ -23,8 +23,8 @@ set_property(TARGET mozillavpn-sources PROPERTY INTERFACE_INCLUDE_DIRECTORIES
 # VPN Client source files
 target_sources(mozillavpn-sources INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/accessiblenotification.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/featurelistcallback.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/featurelist.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/feature/featurelistcallback.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/feature/featurelist.h
     ${CMAKE_CURRENT_SOURCE_DIR}/appimageprovider.h
     ${CMAKE_CURRENT_SOURCE_DIR}/applistprovider.h
     ${CMAKE_CURRENT_SOURCE_DIR}/apppermission.cpp
