+++ src/connectionbenchmark/connectionbenchmark.cpp
@@ -9,7 +9,7 @@
 #include "connectionhealth.h"
 #include "constants.h"
 #include "controller.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "glean/generated/metrics.h"
 #include "leakdetector.h"
 #include "logger.h"
