+++ src/glean/mzglean.cpp
@@ -5,7 +5,7 @@
 #include "mzglean.h"
 
 #include "constants.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "glean/generated/metrics.h"
 #include "glean/generated/pings.h"
 #include "leakdetector.h"
