+++ src/keyregenerator.cpp
@@ -6,7 +6,7 @@
 
 #include "constants.h"
 #include "controller.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "leakdetector.h"
 #include "logger.h"
 #include "mfbt/checkedint.h"
