+++ src/models/servercountrymodel.cpp
@@ -11,7 +11,7 @@
 
 #include "collator.h"
 #include "constants.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "leakdetector.h"
 #include "logger.h"
 #include "mozillavpn.h"
