+++ src/mozillavpn.cpp
@@ -12,7 +12,7 @@
 #include "constants.h"
 #include "controller.h"
 #include "dnshelper.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "frontend/navigationbarmodel.h"
 #include "frontend/navigator.h"
 #include "glean/generated/metrics.h"
