+++ src/qmlengineholder.cpp
@@ -12,6 +12,7 @@
 #include "authenticationinapp/authenticationinapp.h"
 #include "env.h"
 #include "errorhandler.h"
+#include "feature/featuremodel.h"
 #include "frontend/navigationbarmodel.h"
 #include "frontend/navigator.h"
 #include "glean/generated/metrics.h"
@@ -20,7 +21,6 @@
 #include "leakdetector.h"
 #include "localizer.h"
 #include "loghandler.h"
-#include "models/featuremodel.h"
 #include "models/licensemodel.h"
 #include "settingsholder.h"
 #include "theme.h"
