+++ src/sentry/sentryadapter.cpp
@@ -15,7 +15,7 @@
 #include <QStandardPaths>
 
 #include "constants.h"
-#include "feature.h"
+#include "feature/feature.h"
 #include "leakdetector.h"
 #include "logger.h"
 #include "loghandler.h"
