+++ tests/unit/testaddon.cpp
@@ -20,12 +20,12 @@
 #include "addons/conditionwatchers/addonconditionwatchertimestart.h"
 #include "addons/conditionwatchers/addonconditionwatchertriggertimesecs.h"
 #include "addons/manager/addonmanager.h"
-#include "feature.h"
+#include "feature/feature.h"
+#include "feature/featuremodel.h"
 #include "glean/generated/metrics.h"
 #include "glean/mzglean.h"
 #include "helper.h"
 #include "localizer.h"
-#include "models/featuremodel.h"
 #include "qmlengineholder.h"
 #include "qtglean.h"
 #include "settingsholder.h"
