+++ tests/unit_tests/feature/testfeaturemodel.cpp
@@ -5,9 +5,9 @@
 #include "testfeaturemodel.h"
 
 #include "adjust/adjustfiltering.h"
-#include "feature.h"
+#include "feature/feature.h"
+#include "feature/featuremodel.h"
 #include "helper.h"
-#include "models/featuremodel.h"
 #include "settingsholder.h"
 
 void TestFeatureModel::flipOnOff() {
