+++ tests/unit_tests/testaddon.cpp
@@ -19,12 +19,12 @@
 #include "addons/conditionwatchers/addonconditionwatchertimeend.h"
 #include "addons/conditionwatchers/addonconditionwatchertimestart.h"
 #include "addons/conditionwatchers/addonconditionwatchertriggertimesecs.h"
-#include "feature.h"
+#include "feature/feature.h"
+#include "feature/featuremodel.h"
 #include "glean/generated/metrics.h"
 #include "glean/mzglean.h"
 #include "helper.h"
 #include "localizer.h"
-#include "models/featuremodel.h"
 #include "qmlengineholder.h"
 #include "qtglean.h"
 #include "settings/settingsmanager.h"
